Move all components to the components directory
