+++ src/components/App.js
@@ -1,5 +1,5 @@
-import './App.css';
-import './Rule.css';
+import '../App.css';
+import '../Rule.css';
 import React, { PureComponent } from 'react';
 import Query from './Query';
