+++ src/components/DeleteRule.js
