+++ src/components/Rule.js
