+++ src/index.js
@@ -3,6 +3,6 @@ import './index.css';
 import React from 'react';
 import { render } from 'react-dom';
 
-import App from './App';
+import App from './components/App';
 
 render(<App />, document.querySelector('#app'));
